+++ src/SilverStripe/InvalidationExtension.php
@@ -78,22 +78,75 @@ class InvalidationExtension extends Extension
             'action' => $action
         );
 
+        $logger = $this->cache->getLogger();
+
+        // TODO: A bit complex. Needs refactor
         foreach ($this->cache->getConfig() as $name => $inst) {
-            if (isset($inst['invalidation_rules'])) {
-                foreach ($inst['invalidation_rules'] as $rule) {
-                    if ($this->expressionLanguage->evaluate($rule, $vars)) {
-                        $this->cache->flushByName($name);
-                        if ($logger = $this->cache->getLogger()) {
-                            $logger->info(sprintf(
-                                "Cache '%s' invalidated by rule '%s'",
-                                $name,
-                                $rule
-                            ));
-                        }
+            $hasContainsRules = isset($inst['contains']) && is_array($inst['contains']);
+            $hasInvalidationRules = isset($inst['invalidation_rules']) && is_array($inst['invalidation_rules']);
+            
+            // Check to see if there are contains rules, and if there is
+            // we want to ensure that the invalidation rules are only processed
+            // if the class at least passes the invalidation rules
+            if ($hasContainsRules) {
+                $contains = false;
+                foreach ($inst['contains'] as $contain) {
+                    if ($this->owner instanceof $contain) {
+                        $contains = true;
                         break;
                     }
                 }
+                
+                // We don't want to do any more processing if the contains
+                // rules didn't even pass
+                if (!$contains) {
+                    continue;
+                }
+            } else {
+                $contains = true;
             }
+            
+            if ($hasInvalidationRules) {
+                foreach ($inst['invalidation_rules'] as $rule) {
+                    if ($this->expressionLanguage->evaluate($rule, $vars) && $contains) {
+                        $this->invalidate(
+                            $name,
+                            sprintf(
+                                "Cache '%s' invalidated by rule '%s'",
+                                $name,
+                                $rule
+                            ),
+                            $logger
+                        );
+                        break;
+                    }
+                }
+            } elseif ($hasContainsRules && $contains) {
+                // If there aren't any invalidation rules then invalidate the cache
+                // simply based on the contains rules passing
+                $this->invalidate(
+                    $name,
+                    sprintf(
+                        "Cache '%s' invalidated contains rules",
+                        $name
+                    ),
+                    $logger
+                );
+            }
+        }
+    }
+
+    /**
+     * Invalidates a cache by a certain name, and logs if available
+     * @param $name
+     * @param $message
+     * @param null $logger
+     */
+    protected function invalidate($name, $message, $logger = null)
+    {
+        $this->cache->flushByName($name);
+        if ($logger) {
+            $logger->info($message);
         }
     }